create "internal" package, move XmlUtils there
