+++ src/main/java/com/kdgregory/geoutil/lib/gpx/Track.java
@@ -25,7 +25,7 @@ import org.w3c.dom.Element;
 
 import net.sf.practicalxml.DomUtil;
 
-import com.kdgregory.geoutil.lib.shared.XmlUtils;
+import com.kdgregory.geoutil.lib.internal.XmlUtils;
 
 
 /**
