+++ src/main/java/com/kdgregory/geoutil/lib/internal/XmlUtils.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.kdgregory.geoutil.lib.shared;
+package com.kdgregory.geoutil.lib.internal;
 
 import java.util.HashMap;
 import java.util.List;
